Change to initial red LED state
In this revision of the code we changed the way the red LED blinks. In the original code the red LED was turned on and off using delays in the initial state on power on. This revision the LED is turned off when toggle1 is high and turned on when toggle1 is low. This ensures the LED is toggled on and off every second on power on.